+++ study/src/main/java/reflection/servlet/com/example/CharacterEncodingFilter.java
@@ -1,4 +1,4 @@
-package servlet.com.example;
+package reflection.servlet.com.example;
 
 import jakarta.servlet.*;
 import jakarta.servlet.annotation.WebFilter;
@@ -11,6 +11,7 @@ public class CharacterEncodingFilter implements Filter {
     @Override
     public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
         request.getServletContext().log("doFilter() 호출");
+        response.setCharacterEncoding("UTF-8");
         chain.doFilter(request, response);
     }
 }
